code cleanup in init_ref_frame_space()

+++ av1/encoder/block.h
@@ -478,7 +478,7 @@ struct macroblock {
   // The type of mv cost used during motion search
   MV_COST_TYPE mv_cost_type;
 
-  int search_ref_frame[REF_FRAMES];
+  uint8_t search_ref_frame[REF_FRAMES];
 
 #if CONFIG_AV1_HIGHBITDEPTH
   void (*fwd_txfm4x4)(const int16_t *input, tran_low_t *output, int stride);
